refactor display: - move it to a separate namespace - add IDisplay interface
